refactor: reorganize TypeScript definitions for Cloudflare environment variables
